refactor: 도메인별 entity 분리
