+++ src/entities/banner/index.ts
@@ -1,0 +1,3 @@
+export { addBannerApi } from './api/bannerApi';
+export { useAddBanner } from './api/queries';
+export type { AddBannerRequest } from './model/types';
