+++ src/entities/banner/model/types.ts
@@ -1,0 +1,4 @@
+export interface AddBannerRequest {
+  link: string;
+  image: File;
+}
