+++ src/entities/place/model/types.ts
@@ -118,34 +118,3 @@ export interface CreatePlaceRequest extends Location {
 export interface UpdateLocationRequest extends Location {
   placeId: number;
 }
-
-// 로그인 요청
-export interface LoginRequest {
-  loginId: string;
-  password: string;
-}
-
-export interface LoginResponse {
-  tokenResponse: {
-    accessToken: string;
-    refreshToken: string;
-    accessExpireIn: number;
-    refreshExpireIn: number;
-    isFirstLogin: boolean;
-  };
-  userResponse: {
-    id: number;
-    oauthId: string;
-    loginId: string;
-    email: string;
-    nickname: string;
-    studentId: string;
-    imageUrl: string;
-    departmentResponse: string[];
-  };
-}
-
-export interface AddBannerRequest {
-  link: string;
-  image: File;
-}
